+++ R/biographR-package.R
@@ -2,12 +2,15 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom dplyr left_join
 #' @importFrom httr2 req_body_json
 #' @importFrom httr2 req_headers
 #' @importFrom httr2 req_perform
 #' @importFrom httr2 request
 #' @importFrom httr2 resp_body_json
 #' @importFrom jsonlite fromJSON
+#' @importFrom rlang !!
+#' @importFrom rlang sym
 #' @importFrom tibble as_tibble
 #' @importFrom tidyr unite
 ## usethis namespace: end