+++ NEWS.md
@@ -1,3 +1,8 @@
 # biographR (development version)
 
-* Initial CRAN submission.
+* Initial Github package.
+
+# biographR 0.0.0.9001
+
+* Added function clean_columns() to post-process output from get_bio() using ChatGPT
+* Updated vignette for get_bio() function